chore(rust): clippy clean up
