+++ implementations/rust/node/src/node.rs
@@ -1,8 +1,7 @@
 use ockam_common::commands::ockam_commands::*;
-#[allow(unused)]
-use ockam_message::message::*;
-use ockam_router::router::*;
-use ockam_transport::transport::*;
+use ockam_message::*;
+use ockam_router::*;
+use ockam_transport::*;
 use std::str;
 use std::thread::sleep;
 use std::{thread, time};
