+++ implementations/rust/router/Cargo.toml
@@ -5,8 +5,7 @@ authors = ["Ockam Developers"]
 edition = "2018"
 
 [lib]
-crate-type = ["staticlib", "rlib", "cdylib"]
-path = "src/router.rs"
+crate-type = ["rlib"]
 
 [profile.release]
 lto = true
