+++ implementations/rust/router/src/lib.rs
@@ -1,92 +1,80 @@
-#![allow(unused)]
-pub mod router {
-    use ockam_common::commands::ockam_commands::*;
-    use ockam_message::message::*;
-    use std::sync::mpsc::channel;
-    use std::sync::{Arc, Mutex};
-    use std::{thread, time};
+use ockam_common::commands::ockam_commands::*;
+use ockam_message::*;
 
-    pub struct Router {
-        registry: Vec<Option<std::sync::mpsc::Sender<OckamCommand>>>,
-        rx: std::sync::mpsc::Receiver<OckamCommand>,
+pub struct Router {
+    registry: Vec<Option<std::sync::mpsc::Sender<OckamCommand>>>,
+    rx: std::sync::mpsc::Receiver<OckamCommand>,
+}
+
+pub enum MessageDirection {
+    Send,
+    Receive,
+}
+
+impl Router {
+    pub fn new(rx: std::sync::mpsc::Receiver<OckamCommand>) -> Router {
+        Router {
+            registry: vec![Option::None; 256],
+            rx,
+        }
     }
 
-    pub enum MessageDirection {
-        Send,
-        Receive,
+    pub fn poll(&mut self) -> bool {
+        let keep_going;
+        loop {
+            if let Ok(rc) = self.rx.try_recv() {
+                match rc {
+                    OckamCommand::Router(RouterCommand::Stop) => {
+                        println!("quit!");
+                        keep_going = false;
+                        break;
+                    }
+                    OckamCommand::Router(RouterCommand::Register(a_type, tx)) => {
+                        println!("Registering");
+                        self.registry[a_type as usize] = Option::Some(tx);
+                    }
+                    OckamCommand::Router(RouterCommand::ReceiveMessage(m)) => {
+                        if let Err(e) = self.route(m, MessageDirection::Receive) {
+                            println!("Router error: {}", e);
+                        }
+                    }
+                    OckamCommand::Router(RouterCommand::SendMessage(m)) => {
+                        println!("Routing");
+                        if let Err(e) = self.route(m, MessageDirection::Send) {
+                            println!("Router error: {}", e);
+                        }
+                    }
+                    _ => println!("Router received bad command"),
+                }
+            }
+        }
+        keep_going
     }
 
-    impl Router {
-        pub fn new(rx: std::sync::mpsc::Receiver<OckamCommand>) -> Router {
-            Router {
-                registry: vec![Option::None; 256],
-                rx,
-            }
+    fn route(&mut self, m: Message, direction: MessageDirection) -> Result<(), String> {
+        if m.onward_route.addresses.is_empty() {
+            return Err("no route supplied".to_string());
         }
 
-        pub fn poll(&mut self) -> bool {
-            let mut keep_going = true;
-            let mut got = true;
-            while got {
-                got = false;
-                match self.rx.try_recv() {
-                    Ok(rc) => match rc {
-                        OckamCommand::Router(RouterCommand::Stop) => {
-                            println!("quit!");
-                            got = true;
-                            keep_going = false;
-                            break;
-                        }
-                        OckamCommand::Router(RouterCommand::Register(a_type, tx)) => {
-                            println!("Registering");
-                            got = true;
-                            self.registry[a_type as usize] = Option::Some(tx);
-                        }
-                        OckamCommand::Router(RouterCommand::ReceiveMessage(m)) => {
-                            got = true;
-                            self.route(m, MessageDirection::Receive);
-                        }
-                        OckamCommand::Router(RouterCommand::SendMessage(m)) => {
-                            println!("Routing");
-                            got = true;
-                            self.route(m, MessageDirection::Send);
-                        }
-                        _ => println!("Router received bad command"),
-                    },
-                    Err(e) => {}
-                }
-            }
-            keep_going
-        }
-
-        fn route(&mut self, m: Message, direction: MessageDirection) -> Result<(), String> {
-            if m.onward_route.addresses.is_empty() {
-                return Err("no route supplied".to_string());
-            }
-
-            let destination_address = m.onward_route.addresses[0];
-            let address_type = destination_address.a_type;
-            let handler_tx = match &self.registry[address_type as usize] {
-                Some(a) => a,
-                None => return Err("no handler".to_string()),
-            };
-            match address_type {
-                AddressType::Channel => match direction {
-                    MessageDirection::Receive => {
-                        handler_tx.send(OckamCommand::Channel(ChannelCommand::ReceiveMessage(m)));
-                        Ok(())
-                    }
-                    MessageDirection::Send => {
-                        handler_tx.send(OckamCommand::Channel(ChannelCommand::SendMessage(m)));
-                        Ok(())
-                    }
-                },
-                AddressType::Udp => {
-                    handler_tx.send(OckamCommand::Transport(TransportCommand::SendMessage(m)));
-                    Ok(())
-                }
-                _ => Err("not implemented".to_string()),
-            }
+        let destination_address = m.onward_route.addresses[0];
+        let address_type = destination_address.a_type;
+        let handler_tx = match &self.registry[address_type as usize] {
+            Some(a) => a,
+            None => return Err("no handler".to_string()),
+        };
+        match address_type {
+            AddressType::Channel => match direction {
+                MessageDirection::Receive => handler_tx
+                    .send(OckamCommand::Channel(ChannelCommand::ReceiveMessage(m)))
+                    .map_err(|e| e.to_string()),
+                MessageDirection::Send => handler_tx
+                    .send(OckamCommand::Channel(ChannelCommand::SendMessage(m)))
+                    .map_err(|e| e.to_string()),
+            },
+            AddressType::Udp => handler_tx
+                .send(OckamCommand::Transport(TransportCommand::SendMessage(m)))
+                .map_err(|e| e.to_string()),
+            _ => Err("not implemented".to_string()),
         }
     }
 }
